+++ awspricing/__init__.py
@@ -1,25 +1,20 @@
-import requests
-from requests.adapters import HTTPAdapter
-from typing import Dict, Type  # noqa
+from typing import Dict, Type, List
+import boto3
+import json
+import datetime
 
 from .offers import AWSOffer, get_offer_class  # noqa
 from .cache import maybe_read_from_cache, maybe_write_to_cache
 
 
-__version__ = "1.1.5"
+__version__ = "2.0.0"
 
-
-session = requests.Session()
-session.mount('http://', HTTPAdapter(max_retries=5))
-session.mount('https://', HTTPAdapter(max_retries=5))
-
-
-OFFER_BASE_URL = 'https://pricing.us-east-1.amazonaws.com'
-OFFER_INDEX_ENDPOINT = '/offers/v1.0/aws/index.json'
-
-
-_OFFERS = None
 _SERVICES = {}  # type: Dict[str, Type[AWSOffer]]
+service_list = []  # type: List[str]
+
+client = boto3.client('pricing', region_name='us-east-1')
+
+TIME_FORMAT = '%b_%y_%Z'  # month, year, and timezone
 
 
 def _fetch_offers():
@@ -28,69 +23,49 @@ def _fetch_offers():
     if offers is not None:
         return offers
 
-    resp = session.get(OFFER_BASE_URL + OFFER_INDEX_ENDPOINT)
-    resp.raise_for_status()
-    offers = resp.json()['offers']
-
     maybe_write_to_cache(cache_key, offers)
     return offers
 
 
-def _fetch_available_versions(offers, offer_name):
-    cache_key = 'offers_version_{}'.format(offer_name)
-    offers_version = maybe_read_from_cache(cache_key)
-    if offers_version is not None:
-        return offers_version
-
-    offers_version_endpoint = offers[offer_name]['versionIndexUrl']
-    resp = session.get(OFFER_BASE_URL + offers_version_endpoint)
-    resp.raise_for_status()
-    offers_version = resp.json()['versions']
-
-    maybe_write_to_cache(cache_key, offers_version)
-    return offers_version
+def _get_services():
+    global service_list
+    if not service_list:
+        service_list = all_services_names()
+    return service_list
 
 
-def _get_offers():
-    global _OFFERS
-    if not _OFFERS:
-        _OFFERS = _fetch_offers()
-    return _OFFERS
-
-
-def _fetch_offer(offer_name, version='current'):
-    offers = _get_offers()
-    if offer_name not in offers:
-        raise ValueError('Unknown offer name: {}'.format(offer_name))
+def _fetch_offer(offer_name, version=None):
+    services = _get_services()
+    if offer_name not in services:
+        raise ValueError('Unknown offer name, no corresponding AWS Service: {}'.format(offer_name))
+    if not version:
+        version = datetime.datetime.utcnow().strftime(TIME_FORMAT)
 
     cache_key = 'offer_{}_{}'.format(offer_name, version)
     offer = maybe_read_from_cache(cache_key)
     if offer is not None:
         return offer
 
-    if version == 'current':
-        offer_endpoint = offers[offer_name]['currentVersionUrl']
-    else:
-        other_versions = _fetch_available_versions(offers, offer_name)
-        if version in other_versions:
-            offer_endpoint = other_versions[version]['offerVersionUrl']
-        else:
-            raise ValueError('Invalid version specified {}. Must be one of: [{}]'.format(
-                version, ', '.join(sorted(other_versions.keys(), reverse=True))))
-
-    resp = session.get(OFFER_BASE_URL + offer_endpoint)
-    resp.raise_for_status()
-    offer = resp.json()
+    paginator = client.get_paginator('get_products')
+    resp_pages = paginator.paginate(ServiceCode=offer_name, FormatVersion='aws_v1')
+    offer = {}
+    for page in resp_pages:
+        for product in page['PriceList']:
+            product_offer = json.loads(product)
+            sku = product_offer['product']['sku']
+            offer[sku] = product_offer
 
     maybe_write_to_cache(cache_key, offer)
     return offer
 
 
-def all_service_names():
-    return _get_offers().keys()
+def all_services_names():
+    resp = client.describe_services()
+    services = [x['ServiceCode'] for x in resp['Services']]
+    return services
 
 
-def offer(service_name, version='current'):
+def offer(service_name, version=None):
     if service_name not in _SERVICES:
         offer_data = _fetch_offer(service_name, version=version)
         _SERVICES[service_name] = get_offer_class(service_name)(offer_data)