+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0
+
+ * Migrate to use AWS Price List query API instead of offer files
+ * Uses boto3 to access AWS API
+ * Pricing data now pulled lazily by AWS product
+
 ## 1.1.5
 
  * Add capacitystatus attribute for EC2 (#29)